# openapi.yaml
openapi: 3.1.0
info:
  title: TIC API
  version: "0.1.0-mvp"
  description: >
    Contract for the TIC MVP backend. Enum values in this file must match
    shared/constants.json exactly — see docs/CONTRACT_CHANGE_PROCESS.md
    before editing either file.
servers:
  - url: /api/v1
paths:
...
  /admin/employees:
    post:
      summary: Admin/manager creates an employee in their own organization
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, department]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
                department: { type: string }
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  name: { type: string }
                  email: { type: string }
                  department: { type: string }
                  role: { type: string, enum: [EMPLOYEE] }
        "409":
          description: Email already registered
...
  /tasks:
    post:
      summary: Create a task
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string, enum: [TODO] }
...
  /manager/team:
    get:
      summary: Team activity overview, derived from real collected activity data
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items: { $ref: "#/components/schemas/TeamOverviewEntry" }
  /manager/employees/{employee_id}:
    get:
      summary: Single employee detail view for a manager, optionally filtered to one UTC calendar day
      security: [{ bearerAuth: [] }]
      parameters:
        - name: employee_id
          in: path
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
          description: Filter to a single UTC calendar day. Omit for all-time.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeDetail" }
        "404":
          description: Employee not found (or belongs to a different organization)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EmployeeDashboard:
      type: object
      description: >
        Reconciled with the actually-implemented response shape of
        GET /employee/me/dashboard and GET /manager/employees/{id} (previous
        versions of this schema described aggregate fields — active_time,
        workload_percentage, etc. — that the endpoint never actually
        returned; see docs/CONTRACT_CHANGE_PROCESS.md).
      properties:
        workPulse:
          type: array
          items:
            type: object
            properties:
              app: { type: string }
              start: { type: string }
              end: { type: string }
              duration: { type: string }
              type: { type: string }
              width: { type: string }
        applications:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              duration: { type: string }
              category: { type: string }
              lastActive: { type: string }
        hourlyActivity:
          type: array
          items:
            type: object
            properties:
              hour: { type: string }
              intensityPercent: { type: number }
        activeSeconds: { type: integer }
        idleSeconds: { type: integer }
        sessionCount: { type: integer }
...
    TaskCreateRequest:
      type: object
      required: [title, estimated_hours]
      properties:
        title: { type: string }
        description: { type: string }
        assigned_to: { type: [integer, "null"] }
        estimated_hours: { type: number }
        complexity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        deadline: { type: string, format: date }
...
    TeamOverviewEntry:
      type: object
      description: >
        Real, derived-from-activity-data team summary. Does NOT include
        task-based workload_percentage — that requires task estimated_hours
        vs. capacity, which is out of scope for this pass.
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        department: { type: string }
        activeTime: { type: string }
        sessionCount: { type: integer }
        topApp: { type: [string, "null"] }
        lastActive: { type: string }
    EmployeeDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        department: { type: string }
        status: { type: string }
        dashboard: { $ref: "#/components/schemas/EmployeeDashboard" }
...